add inspect_artifacts command
