+++ recipe/meta.yaml
@@ -79,6 +79,7 @@ requirements:
     - git     # [unix]
     - libarchive
     - conda-forge-metadata >=0.4.1
+    - conda-package-handling >=2.3.0
   run_constrained:
     - boa >=0.8,<1.0
